+++ gin/jwtx/s.Logout.go
@@ -13,13 +13,13 @@ import (
 //	c             *gin.Context
 //	loginGroup    string       登录的 token 分组
 //	loginTerminal string       登录的 终端名称（当传值为“all”时，无视配置，强制登出所有终端）
-//	accountID     uint32       登录的 账户 ID
+//	accountID     uint64       登录的 账户 ID
 //
 // e.g.
 //
 //	jwtx.Singleton.Logout(c, "admin", "pc", 1)
 //	jwtx.Singleton.Logout(c, "admin", "all", 1)
-func (s *SingletonT) Logout(c *gin.Context, loginGroup string, loginTerminal string, accountID uint32) (err error) {
+func (s *SingletonT) Logout(c *gin.Context, loginGroup string, loginTerminal string, accountID uint64) (err error) {
 	// 获取分组配置
 	var (
 		q      = query.Use(s.DB[loginGroup])
